+++ core/platform/k8s/src/main/kotlin/org/http4k/cloudnative/deprecatedCloudNative.kt
@@ -1,0 +1,4 @@
+package org.http4k.cloudnative
+
+@Deprecated("Use org.http4k.k8s.Http4kK8sServer instead", ReplaceWith("org.http4k.k8s.Http4kK8sServer"))
+typealias Http4kK8sServer = org.http4k.k8s.Http4kK8sServer