+++ core/platform/k8s/src/main/kotlin/org/http4k/server/deprecatedCloudNativeExt.kt
@@ -1,0 +1,7 @@
+package org.http4k.server
+
+import org.http4k.config.Port
+import org.http4k.core.HttpHandler
+
+@Deprecated("This is extraneous", ReplaceWith("asServer(fn(8000))"))
+fun HttpHandler.asServer(fn: (Int) -> ServerConfig, port: Port) = fn(port.value).toServer(this)